django/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-ninja==0.13.0
pycodestyle==2.7.0
pydantic==1.8.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
django-cors-headers